# Optimizations for ARM64