@FiloSottile 2024-05-24
Nice! Always nice getting pure Go optimizations.

Since these are parts of the library that track upstream, would you consider mailing them as Go CLs? If not, have you signed the Google CLA and are ok with me submitting them on your behalf after we merge them here, to keep upstream in sync?