@egonelbre 2024-05-24
Sure, I can send them as Go CLs. And yeah, I have signed the CLA.