@egonelbre 2024-05-24
PS: I'll try to add *19 rule first, if I'm making it against Go, so one of the CL-s wouldn't be necessary -- i.e. https://github.com/golang/go/issues/67575