@egonelbre 2025-02-18
Unfortunately the general `* 19` optimization turned out to be more complicated. The performance varies based on the device and context. Anyways, submitted the CL-s with manual `*19` optimization https://go-review.googlesource.com/c/go/+/650277